--- routes/user.routes.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const UsersController = require("../controllers/user.controller");
-
-const usersController = new UsersController();
-
-router.post("/", usersController.searchUser);
-
-module.exports = router;
